changed all pointer casts to explicit static_cast or reinterpret_cast. Added more tests for forward lookahead asserts.
git-svn-id: http://svn.boost.org/svn/boost/trunk@12575 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/libs/regex/test/regress/tests.cpp
@@ -365,7 +365,7 @@ void cpp_tests(const reg_expression<C, T, A>& e, bool recurse = true)
                }
             }
          }
-         if((false == recurse) && (matches[0] == 0) && (matches[1] == search_text.size()))
+         if((false == recurse) && (matches[0] == 0) && (matches[1] == static_cast<int>(search_text.size())))
          {
             //
             // match expected on whole string, so all versions
